addons: nanokvm-pro: add more pikvm cleanup

+++ scripts/addons/nanokvm-pro/addon.mk
@@ -276,19 +276,28 @@ $(BUILDDIR)/nanokvm-pro-firmware-stamp: $(BUILDDIR)/aic8800-firmware-stamp $(BUI
 $(BUILDDIR)/nanokvm-pro-pikvm-stamp: $(BUILDDIR)/nanokvm-pro-prepare-stamp $(BUILDDIR)/pikvm-stamp
 	@cd $(BUILDDIR)/nanokvm-pro ; dpkg-deb -R nanokvm_pro_$(NANOKVM_PRO_VERSION)/pikvm_$(NANOKVM_PRO_VERSION)_$(DEB_ARCH).deb $(NANOKVM_PRO_PIKVM_PACKAGE_DIR)
 	@mkdir -p $(NANOKVM_PRO_PIKVM_PACKAGE_DIR)/usr/local/include/
-	@cd $(NANOKVM_PRO_PIKVM_PACKAGE_DIR) && mv usr/include/gpiod.h usr/local/include/
+	@cd $(NANOKVM_PRO_PIKVM_PACKAGE_DIR) && [ ! -e usr/include/gpiod.h ] || mv usr/include/gpiod.h usr/local/include/
 	@if [ -e $(PIKVM_BUILD_DIR)/out ]; then \
 		cd $(NANOKVM_PRO_PIKVM_PACKAGE_DIR) && \
+		mkdir -p etc/kvmd/depends && \
+		mkdir -p etc/kvmd.backup && \
+		mv etc/kvmd/depends etc/kvmd.backup/ && \
 		rm -rf etc/janus/ && \
+		rm -rf etc/kvmd/ && \
 		rm -rf etc/sudoers.d/ && \
 		rm -rf usr/include/ && \
 		rm -f usr/bin/gpio* && \
+		rm -f usr/bin/janus* && \
+		rm -f usr/bin/kvmd* && \
+		rm -f usr/bin/nanokvm* && \
+		rm -f usr/bin/vcgen* && \
 		rm -rf usr/lib/ && \
 		rm -rf usr/share/ && \
 		rm -rf usr/local/include/ && \
 		rm -rf usr/local/lib/ && \
 		rm -rf usr/local/share/ && \
 		rm -rf var/lib/ && \
+		mv etc/kvmd.backup etc/kvmd && \
 		rsync -avpPxH $(PIKVM_BUILD_DIR)/out/ ./ && \
 		echo $(NANOKVM_PRO_VERSION) > etc/kvmd/version ; \
 	fi
